pyproject.toml
[project]
name = "fcpm-simulation"
version = "2.0.0"
description = "FCPM Simulation and Reconstruction for Liquid Crystal Director Fields"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "FCPM Simulation Team" },
]
keywords = [
    "FCPM",
    "liquid crystal",
    "director field",
    "microscopy",
    "simulation",
    "reconstruction",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Visualization",
]
dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
]

[project.optional-dependencies]
full = [
    "pymaxflow>=1.2.0",
    "tifffile>=2021.0.0",
    "h5py>=3.8.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
]
perf = [
    "numba>=0.57.0",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.0",
    "mkdocstrings[python]>=0.24",
]

[project.scripts]
fcpm = "fcpm.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["fcpm"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.ruff]
target-version = "py310"
line-length = 100
